Adjusted the unit tests for morphology helper files
